+++ resources/views/layouts/app.blade.php
@@ -33,14 +33,14 @@
                 <!-- Page Header -->
                 @isset($header)
                     <header class="bg-gray-900 border-b border-gray-800">
-                        <div class="max-w-7xl mx-auto py-4 px-4 sm:px-6 lg:px-8">
+                        <div class="w-full py-4 px-4 sm:px-6 lg:px-8">
                             {{ $header }}
                         </div>
                     </header>
                 @endisset
 
                 <!-- Page Content -->
-                <main class="flex-1 p-4 sm:p-6 lg:p-8">
+                <main class="flex-1 p-4 sm:p-6 lg:p-8 w-full">
                     {{ $slot }}
                 </main>
             </div>