+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\Web\DashboardController;
 use App\Http\Controllers\Web\DeploymentController;
+use App\Http\Controllers\Web\LogController;
 use App\Http\Controllers\Web\ProjectController;
 use App\Http\Controllers\Web\QueueController;
 use App\Http\Controllers\Web\ServiceController;
@@ -28,6 +29,8 @@ Route::middleware('auth')->group(function () {
     Route::post('projects/{project}/services/{service}/ping', [ServiceController::class, 'ping'])->name('projects.services.ping');
 
     Route::resource('projects.deployments', DeploymentController::class)->only(['index', 'show']);
+
+    Route::get('projects/{project}/logs', [LogController::class, 'index'])->name('projects.logs.index');
 
     Route::get('/queues', QueueController::class)->name('queues.index');
 });